isolate the source_gen dependency form the stacked_locator_content_generator
