+++ packages/stacked_generator/lib/src/generators/exceptions/invalid_generator_input_exception.dart
@@ -1,0 +1,12 @@
+import 'package:analyzer/dart/element/element.dart';
+
+class InvalidGeneratorInputException implements Exception {
+  final String message;
+  final String? todo;
+  final Element? element;
+  const InvalidGeneratorInputException(this.message, {this.todo, this.element});
+
+  @override
+  String toString() =>
+      'InvalidGeneratorInputException: message: $message\n todo: $todo\n element: $element';
+}
