+++ packages/stacked_generator/lib/src/generators/getit/stacked_locator_parameter_resolver.dart
@@ -3,30 +3,9 @@ import 'package:source_gen/source_gen.dart';
 import 'package:stacked_core/stacked_core.dart';
 import 'package:stacked_generator/import_resolver.dart';
 import 'package:stacked_generator/utils.dart';
+import 'dependency_param_config.dart';
 
 const _factoryParamChecker = TypeChecker.fromRuntime(FactoryParam);
-
-class DependencyParamConfig {
-  final String? type;
-  final String? name;
-  final String? alias;
-  final bool isPositional;
-  final bool isRequired;
-  final String? defaultValueCode;
-  final Set<String>? imports;
-  final bool isFactoryParam;
-
-  DependencyParamConfig({
-    this.type,
-    this.name,
-    this.alias,
-    this.isFactoryParam = false,
-    this.isPositional = false,
-    this.isRequired = false,
-    this.defaultValueCode,
-    this.imports,
-  });
-}
 
 class DependencyParameterResolver {
   final ImportResolver _importResolver;
